map EDBB to EDDB (#20)
As Tegel/EDDT closed on 8 November 2020 and Tegel CTR was revoked,
there are no further active airfields covered by EDBB, only EDDB
is covered from now on. However, APP/DEP/DIR stations will remain
named EDBB instead of EDDB on VATSIM in the near future, so the
mapping is not removed yet but redirected to the correct place.