[TASK] do not center on list models when double clicking
Centering the map on a MapClient can be achieved by using the "show on
map" function from its dialog.

Click also opens that dialog which is then in the middle of the view so
double click did not improve the UX because it did not provide a quicker
path to that user intention.