[TASK] remove platform specific .pro files
It had some merrits while I was developing Linux and Mac simultaneously
out of the same shared source tree. It was a personal convenience
workaround to make Qt Creator for my specific use case happy.

But for newcomers it is just confusing and even error-prone since
creating new project files in Qt Creator would be added to the platform
.pro file.