[FEATURE] release early, release often (#63)
Download artifacts (.zip, .tgz):

This produces runable artifacts for each push and each change to any pull
request.

Pre-releases:

The Github workflow creates Github (pre-)releases for each push to this
repository.
The naming scheme is `v2.3.0-6-g29966c2` meaning "built from commit 29966c2
which is 6 commits behind the main release v2.3.0". Albeit not totally
obvious which repository or branch has built that artifact, it avoids
any ambiguity.

Main releases:

Main releases are automatically created from tagged commits (e.g. v2.3.0)
which are created and pushed to the release branch by maintainers.

Note to contributors:

For the "versiony string" to be the most helpful, contributors should
use `git fetch --tags upstream && git push --tags` before opening a pull
request to have the newest tags in their repository.

This implements most of the wishes in #56.

Caveats:

Due to
https://github.community/t/github-actions-are-severely-limited-on-prs/18179
we can not create releases for PRs nor is it trivial to post a link to the
build artifact into the pull request. I hope the Github UI will improve in
that regard and make it easier to discover the build artifacts from a PR
or an issue that's being affected by a commit.