Remove download/statusLocation and add necessary migration
The previous approach stored download/statusLocation instead of
obtaining the necessary information from download/network.
The previous defaults were a user defined network, however the
statusLocation (Which is not the same as userLocation and never actually
displayed to the user) was already set to the VATSIM URL, causing
QuteScoop to display VATSIM data, but claim to use a user defined
network.

A migration has been added to change download/network to 0 (VATSIM) if
it was previously 1 (user defined) and the statusLocation matches the
old VATSIM status location