[BUGFIX] add workaround for Win/Nvidia mouse coordinates
This is a work around. While the MouseEvent seems to
arbitrarily contain wrong coordinates in `pos()` on some Nvidia drivers
under some circumstances (confirmed in driver versions 496.49 and 511.65),
using the widget's `mapFromGlobal(QCursor::pos())` seems to not carry
the error and be transparent when the error does not surface at all.

Resolves: qutescoop/qutescoop#46