[TASK] remove forced dialog resizing
This fixes dialog sizes changing by themselves, even on consecutive
`refresh()`s with the same data which has lead to bad UX.

Dialogs will still try to grow in size when but not that aggressively.

On the other hand the dialog size might now not accomodate for all the
data shown. So users might occasionally have to resize the dialog
themselves to see long routes, ATISes or flight plan remarks.

Resolves: #124