[FEATURE] match partial sector names
This matches any part of the controller login (from 2 letters upwards)
to a sector.

It deals with VATSIM inconsistencies.

It allows for example to show controller "LOVV-B_CTR", although we do
not know a sector "LOVV-B", but we know "LOVV".

It keeps EURW-N_FSS working (where we have a sector "EURW-N").

This also removes much redundant logic and navdata probing for airport
controllers which improves performance on Whazzup data updates.

Fixes: #159