[FEATURE] show relevant route parts
This allows to use plotted routes better for a short-term situation
preview.

It shows the less relevant future part of a route with more transparency and
with a slimmer line.

The cutoff time is taken from
Preferences -> Display(airports) -> traffic -> arriving in ... hours
(and halved).

The used style for the relevant part of the future route is still
Preferences -> Display (sectors & pilots) -> Pilots -> plot to
destination.

The less relevant part derives from that style, but is
* 20% darker
* 50% less opaque
* half the width (but min. 0.5 pixels)

Fixes: #225
Based on: #226