[FEATURE] Add more airport congestion options - glow and size.
This adds a new visualization ("glow") to airport traffic/congestion. It
can be enabled alone or together with the existing ("ring").

It adds the option to configure radius and color separately for a
minimum and a maximum congestion. For values in between the result is
interpolated and above it is clamped to the given values for maximum.

Planes with status "taxiing in" and "blocked" are now ignored for
the filtered (default for {traffic} label content) congestion numbers.

The configurable default "minimum movements" has been raised to 20 to
account for the higher traffic levels.