[TASK] Simplify logging to stdout (dev)
It was a bit cumbersome to quickly print Qt container types to stdout.
Now we can just use qCritical() << ... . It ends up in the log and on
stdout.

Also qDebug() and friends bring in file, line number and function
signature now (and that is enabled for release builds, too).
In places where we want to customize that we use QMessageLogger (see
Airac.cpp for example).