[TASK] use SOBT/STA terminology
Our previous terminology was misleading.

Also our latest round of reading tea leaves has given us the impression
that VATSIM might probably under circumstances inshallah refer to SOBT
now in the flightplan form. SimBrief uses that field to prefill it with
the SOBT as well.
It is still called `flightplan.deptime` in the status stream though. (?!)